pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "arabic-engine"
version = "0.2.0"
description = "A computational Arabic language engine based on atomic-beginning law and ascending closure."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "sonaiso" },
]
keywords = ["arabic", "nlp", "linguistics", "computational-linguistics"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
]
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.4.0",
]

[project.urls]
Repository = "https://github.com/sonaiso/7-4-26-mahdi"

[tool.setuptools.packages.find]
include = ["arabic_engine*"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
